Lifted state up for color coding, so that it persists across sidebar panel changes
