+++ src/components/AnnotationDesktop/LayersPanel/ColorSettings/index.ts
@@ -1,1 +1,2 @@
 export * from './ColorSettings';
+export * from './ColorState';
