+++ src/components/AnnotationDesktop/LayersPanel/index.ts
@@ -1,1 +1,3 @@
 export * from './LayersPanel';
+export * from './ColorSettings';
+export * from './FilterSettings';
